Changed x3d_freelist_create() to use uint16 for mem_size instead of size_t

+++ src/C/memory/freeslist.c
@@ -35,7 +35,7 @@
 ///
 /// @note mem_size MUST be a multiple of block_size!
 ///////////////////////////////////////////////////////////////////////////////
-void x3d_freelist_create(X3D_FreeList* list, void* mem, size_t mem_size,
+void x3d_freelist_create(X3D_FreeList* list, void* mem, uint16 mem_size,
     int16 block_size, int16 next_pointer_offset, int16 block_id_offset,
     int16 id) {
